+++ lib/openstudio/workflow/multi_delegator.rb
@@ -21,8 +21,6 @@ require 'logger'
 
 class Logger
   def format_message(severity, datetime, _progname, msg)
-    # "#{datetime} (#{$$}) #{msg}\n"
-    # "#{datetime}: #{msg}\n"
     "[%s %s] %s\n" % [datetime.strftime('%H:%M:%S.%6N'), severity, msg]
   end
 end
@@ -43,6 +41,6 @@ class MultiDelegator
   end
 
   class <<self
-    alias_method :to, :new
+    alias to new
   end
 end