+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 OpenStudio::Workflow Change Log
 ==================================
 
+Version '1.0.0-rc1' (Not released)
+-------------------
+* Use specific version of OpenStudio Analysis Gem
+* Fix tests
+* Remove to_underscore method (use OpenStudio.toUnderscoreCase(...) if needed)
+
+Version '1.0.0-alpha.0'
+-----------------------
+* Support new workflow for PAT 2.0. Many of these changes are breaking changes
+* Remove rake from gemspec
+* Update rubyzip to version 1.2
+
 Version 0.2.0
 ------------------
 * Use EnergyPlus that is packaged with OpenStudio. The branch of this repo is no longer tied to the version of EnergyPlus